release: tree-sitter 0.26 upgrade, barrel re-export, full test gap closure
## Tree-sitter 0.24 → 0.26
- Bump tree-sitter core to 0.26.9; tree-sitter-language stays at 0.1.x (ABI compatible)
- 6 grammar version bumps: python 0.25, c 0.24, go 0.25, rust 0.24, php 0.24, bash 0.25
- Unlock c-sharp from =0.23.0 pin
- Fix node.child(i) type change: usize → u32 (tree-sitter 0.26 API)
- Fix StreamingIterator import: streaming_iterator crate → tree_sitter::StreamingIterator re-export
- Remove streaming-iterator crate dependency

## Cangjie ABI compatibility
- tree-sitter 0.26 supports MAX_ABI ≥ 15 → Cangjie (ABI 15) now fully functional
- Remove ABI version workaround in cangjie.rs tests
- Fix references.scm: remove non-existent typeAnnotation node
- Add cangjie to all-languages (both atlas-cli and atlas-engine Cargo.toml)
- Add golden_cangjie_{simple,imports,calls} fixtures with auto-bootstrap expected.json

## Barrel re-export (TypeScript)
- Add ImportKind::ExportFrom enum variant
- Add export capture patterns to imports.scm (export * from, export { X } from)
- Add normalize_ts_export logic + extract_export_module_from_ancestor
- Add resolve_relative_module DB method for path-based module resolution
- Implement follow_reexport_chain with recursive depth=10 + cycle guard
- Wire resolve_through_reexports into ResolutionStrategy::ImportResolved
- Add ts_barrel_reexport_chain_resolves_to_source integration test

## Dataflow cascade delete
- Add find_dataflow_edges_by_file DB method with JOIN query
- Add ts_delete_file_cascades_dataflow integration test

## Test gap closure (8/8 resolved)
- Bash capability: test_bash_capability_is_symbolic
- Cangjie capability: test_experimental_languages_in_all_compiled
- Java trace: fix over-strict partial_result assertion (dataflow is supported_with_limitations)
- All 26 trace CLI E2E + 33 trace MCP E2E pass

## Documentation
- Renumber docs to sequential 01-07 (no gaps)
- Archive and delete all stale docs (architecture.md, archive/*)
- Merge interprocedural summary design into 04-roadmap.md
- Add authoritative capability table to 03-current-architecture.md
- Add tree-sitter dataflow pipeline technical explainer to README.md
- Fix 7 doc cross-reference mismatches (stale links, wrong function names)
- Update 05-testing-spec.md §6: all 8 gaps now resolved

## Test results
- extraction: 160 passed
- golden fixtures: 36 passed
- integration: 20 passed
- trace CLI E2E: 26 passed
- trace MCP E2E: 33 passed
- types: 80 passed
- db: 15 passed (default features)

+++ crates/atlas-engine/crates/extraction/src/dataflow_builder.rs
@@ -471,7 +471,7 @@ fn walk_for_assign_edges(
                 let right_key = NodePosKey { start_byte: right_node.start_byte() as u32, end_byte: right_node.end_byte() as u32, kind: DataNodeKind::Expr };
                 if let Some(&source_id) = pos_map.get(&right_key) {
                     for i in 0..left_node.child_count() {
-                        if let Some(child) = left_node.child(i) {
+                        if let Some(child) = left_node.child(i as u32) {
                             if child.is_named() && child.kind() == "identifier" {
                                 let child_key = NodePosKey { start_byte: child.start_byte() as u32, end_byte: child.end_byte() as u32, kind: DataNodeKind::Local };
                                 if let Some(&target_id) = pos_map.get(&child_key) {
@@ -512,7 +512,7 @@ fn walk_for_assign_edges(
 
     // Recurse into children
     for i in 0..node.child_count() {
-        if let Some(child) = node.child(i) {
+        if let Some(child) = node.child(i as u32) {
             walk_for_assign_edges(child, source, pos_map, _all_nodes, edges);
         }
     }
@@ -527,11 +527,11 @@ pub(crate) fn create_assign_edges_from_expression_lists(
     edges: &mut Vec<DataFlowEdge>,
 ) {
     let left_nodes: Vec<tree_sitter::Node> = (0..left_list.child_count())
-        .filter_map(|i| left_list.child(i))
+        .filter_map(|i| left_list.child(i as u32))
         .filter(|c| c.is_named())
         .collect();
     let right_nodes: Vec<tree_sitter::Node> = (0..right_list.child_count())
-        .filter_map(|i| right_list.child(i))
+        .filter_map(|i| right_list.child(i as u32))
         .filter(|c| c.is_named())
         .collect();
     for (i, left_node) in left_nodes.iter().enumerate() {
@@ -556,7 +556,7 @@ pub(crate) fn create_assign_edges_from_expression_lists(
 fn find_identifier_child(node: tree_sitter::Node) -> Option<tree_sitter::Node> {
     if node.kind() == "identifier" { return Some(node); }
     for i in 0..node.child_count() {
-        if let Some(child) = node.child(i) {
+        if let Some(child) = node.child(i as u32) {
             if let Some(found) = find_identifier_child(child) { return Some(found); }
         }
     }
@@ -1102,7 +1102,7 @@ pub(crate) fn collect_pattern_bindings<'a>(pattern_node: tree_sitter::Node<'a>, 
         // Python destructuring: a, b = pair  /  (a, b) = pair
         "pattern_list" | "tuple_pattern" | "list_pattern" => {
             for i in 0..pattern_node.child_count() {
-                if let Some(child) = pattern_node.child(i) {
+                if let Some(child) = pattern_node.child(i as u32) {
                     if child.kind() == "identifier" {
                         out.push(child);
                     } else {
@@ -1113,14 +1113,14 @@ pub(crate) fn collect_pattern_bindings<'a>(pattern_node: tree_sitter::Node<'a>, 
         }
         "object_pattern" | "array_pattern" | "rest_pattern" => {
             for i in 0..pattern_node.child_count() {
-                if let Some(child) = pattern_node.child(i) {
+                if let Some(child) = pattern_node.child(i as u32) {
                     collect_pattern_bindings(child, out);
                 }
             }
         }
         _ => {
             for i in 0..pattern_node.child_count() {
-                if let Some(child) = pattern_node.child(i) {
+                if let Some(child) = pattern_node.child(i as u32) {
                     collect_pattern_bindings(child, out);
                 }
             }
